requirements.txt
flask
bcrypt
flask-sqlalchemy
PyMySQL
pandas
flask_cors
locustio
gunicorn
pylibmc
yahoo-finance
celery
